# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: PZIO Task Management System
  description: Backend REST API for the team-project task management system. See documents/4.pdf (SAD)
    for the canonical API contract.
  version: 0.1.0
paths:
...
  /api/admin/task-types:
    post:
      tags:
      - Admin
      summary: Add a task type (Admin)
      description: Adds a new task type to the system dictionary. Requires Administrator role.
      operationId: add_task_type_api_admin_task_types_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskTypeCreate'
        required: true
      responses:
        '201':
          description: Task type created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskTypeRead'
        '400':
          description: Validation error
        '401':
          description: Missing or invalid token
        '403':
          description: Insufficient privileges
        '409':
          description: Task type with this name already exists
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/task-types:
    get:
      tags:
      - Admin
      summary: Get task type dictionary
      description: Returns the list of task types defined in the system.
      operationId: get_task_types_api_task_types_get
      responses:
        '200':
          description: List of task types
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TaskTypeRead'
                type: array
                title: Response Get Task Types Api Task Types Get
        '401':
          description: Missing or invalid token
      security:
      - HTTPBearer: []
  /api/admin/task-types/{task_type_id}:
    delete:
      tags:
      - Admin
      summary: Remove a task type (Admin)
      description: Removes a task type from the system. Requires Administrator role. Cannot remove the
        last task type.
      operationId: remove_task_type_api_admin_task_types__task_type_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: task_type_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Type Id
      responses:
        '204':
          description: Task type deleted successfully
        '400':
          description: Cannot delete the last task type
        '401':
          description: Missing or invalid token
        '403':
          description: Insufficient privileges
        '404':
          description: Task type not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/backups:
    post:
      tags:
      - Admin
      summary: Force a database backup (Admin)
      description: Triggers a file-level copy of the SQLite database into the configured backup directory.
        Requires Administrator role.
      operationId: force_backup_api_admin_backups_post
      responses:
        '201':
          description: Backup created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupRead'
        '401':
          description: Missing or invalid token
        '403':
          description: Insufficient privileges
        '409':
          description: Backup already in progress or exists
        '500':
          description: Backup failed
      security:
      - HTTPBearer: []
  /api/tasks/{task_id}/history:
    get:
      tags:
      - Admin
      summary: Get task change history
      description: Returns the audit log entries for a task in chronological order.
      operationId: get_task_history_api_tasks__task_id__history_get
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
          title: Task Id
      responses:
        '200':
          description: List of audit entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityLogRead'
                title: Response Get Task History Api Tasks  Task Id  History Get
        '401':
          description: Missing or invalid token
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    ActivityLogRead:
      properties:
        activityLogId:
          type: integer
          title: Activitylogid
        taskId:
          type: integer
          title: Taskid
        userId:
          type: integer
          title: Userid
        action:
          type: string
          title: Action
        fieldName:
          anyOf:
          - type: string
          - type: 'null'
          title: Fieldname
        oldValue:
          anyOf:
          - type: string
          - type: 'null'
          title: Oldvalue
        newValue:
          anyOf:
          - type: string
          - type: 'null'
          title: Newvalue
        createdAt:
          type: string
          format: date-time
          title: Createdat
      type: object
      required:
      - activityLogId
      - taskId
      - userId
      - action
      - createdAt
      title: ActivityLogRead
      description: Single audit entry returned by `GET /api/tasks/{id}/history` (SAD §4.5).
      example:
        action: STATUS_CHANGE
        activityLogId: 87
        createdAt: '2026-05-14T13:15:00Z'
        fieldName: status
        newValue: InProgress
        oldValue: ToDo
        taskId: 123
        userId: 42
...
    BackupRead:
      properties:
        backupId:
          type: integer
          title: Backupid
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        status:
          type: string
          title: Status
      type: object
      required:
      - backupId
      - timestamp
      - status
      title: BackupRead
      description: 'Response for `POST /api/admin/backups` (SAD §4.5).


        PDF specifies the response shape as `{backupId, timestamp, status}` —

        `timestamp` is read from the ORM column `created_at` via `validation_alias`.'
      example:
        backupId: 12
        status: completed
        timestamp: '2026-05-14T03:00:00Z'
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TaskTypeCreate:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
      type: object
      required:
      - name
      title: TaskTypeCreate
      description: Body for `POST /api/admin/task-types` (SAD §4.5).
      example:
        name: Epic
    TaskTypeRead:
      properties:
        taskTypeId:
          type: integer
          title: Tasktypeid
        name:
          type: string
          title: Name
        createdAt:
          type: string
          format: date-time
          title: Createdat
      type: object
      required:
      - taskTypeId
      - name
      - createdAt
      title: TaskTypeRead
      description: Public dictionary entry returned by /api/task-types and /api/admin/task-types.
      example:
        createdAt: '2026-05-14T08:00:00Z'
        name: Epic
        taskTypeId: 3
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...